requirements.txt
streamlit
pandas
numpy
yfinance
vectorbt
python-dotenv
PyYAML
plotly
